+++ PacletInfo.m
@@ -1,29 +1,25 @@
 (* Paclet Info File *)
 
-(* created 2019/11/01*)
+(* created 2020/01/12*)
 
 Paclet[
     Name -> "MaXrd",
-    Version -> "2.0.0",
+    Version -> "2.1.0",
     MathematicaVersion -> "11.3+",
     Description -> "Database and utilities related to crystallography and X-ray scattering.",
     Creator -> "Stian Penev Ramsnes",
     URL -> "https://github.com/Stianpr20/MaXrd",
     Extensions -> 
         {
-            {"Application",
-            	Context -> "MaXrd`"
+            {"Application", Context -> "MaXrd`"}, 
+            {"Kernel", Context -> 
+                {"MaXrd`"}
             }, 
-            {"Kernel",
-            	Context -> {"MaXrd`"}
-            },
-            {"Documentation",
-            	Language -> "English",
-            	MainPage -> "Guides/MaXrd",
-            	Resources -> {"Guides/MaXrd"}
-            }, 
-            {"FrontEnd",
-            	Context -> {"MaXrd"}
+            {"Documentation", Resources -> 
+                {"Guides/MaXrd"}
+            , Language -> "English", MainPage -> "Guides/MaXrd"}, 
+            {"FrontEnd", Context -> 
+                {"MaXrd"}
             }
         }
 ]