+++ MaXrd/Changelog.md
@@ -1,5 +1,19 @@
 # *MaXrd*: Mathematica X-ray diffraction package – change log
 
+## Version 2.1.0
+### New content
+- Added the option `"IncludeStructureSizeInfo"` to `ExportCrystalData`.
+- Added the option `"ScalingFactor"` to `SimulateDiffractionPattern`.
+
+### Improvements and fixes
+- Fixed a bug in `SimulateDiffractionPattern` where the use of *DISCUS* would not work correctly.
+- Created a `UserData` directory in the package root and moved `CrystalData.m` here. `$CrystalData` and `ResetCrystalData` updated to conform with this change.
+- `ImportCrystalData` now returns instead of aborting if user cancels import.
+- Improved the package unit test.
+- `$MaXrdPath` updated with support for *Windows*.
+- Minor documentation updates.
+
+
 ## Version 2.0.0
 ### Changes
 - `InputCheck["DomainRotation"]` has been replaced with `InputCheck["RotationTransformation"]`, which is more versatile (now used in `DomainPlot`, `EmbedStructure` and `SynthesiseStructure`) and uses different rotation options (`"RotationAnchorReference"`, `"RotationAnchorShift"`, `"RotationAxes"`).